Move EVP compat functions to a separate file
